backend/requirements.txt
flask
flask-cors
librosa
speechbrain
torch
torchaudio
numpy
scipy
web3
python-dotenv
fuzzy-extractor
flwr
soundfile
pydub
